Move all circe-specific code to its own project
