+++ circe/src/main/scala/com/rallyhealth/vapors/v1/dsl/circe/HasEncoder.scala
